+++ authority-portal-backend/authority-portal-quarkus/src/main/kotlin/de/sovity/authorityportal/web/utils/idmanagement/ClientIdUtils.kt
@@ -22,7 +22,6 @@ import de.sovity.authorityportal.db.jooq.Tables
 import jakarta.enterprise.context.ApplicationScoped
 import jakarta.inject.Inject
 import org.jooq.DSLContext
-import java.util.Locale
 
 @ApplicationScoped
 class ClientIdUtils {