+++ authority-portal-backend/catalog-crawler/catalog-crawler/src/main/java/de/sovity/edc/utils/config/model/ConfigProp.java
@@ -18,7 +18,11 @@
 
 package de.sovity.edc.utils.config.model;
 
-import lombok.*;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
 import org.eclipse.edc.spi.system.configuration.Config;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;